move bulk add forms to bulk.py and remove redundant bulk_add.py
