+++ netbox_inventory/forms/__init__.py
@@ -1,6 +1,5 @@
 from .assign import *
 from .bulk import *
-from .bulk_add import *
 from .create import *
 from .filters import *
 from .models import *
